requirements.txt
streamlit
opencv-python
numpy
faiss-cpu
pandas
plotly
ultralytics
mediapipe
onnxruntime
insightface